차트 페이지 버그 수정 및 int 타입을 long 타입으로 바꿈.

+++ rest/src/main/java/com/itwill/rest/domain/AlbumLikeId.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor @AllArgsConstructor
 @Builder @Embeddable
 public class AlbumLikeId {
-	private Integer albumId;
+	private Long albumId;
 
-	private Integer id;
+	private Long id;
 
 }
