+++ rest/src/main/java/com/itwill/rest/repository/ArtistRoleRepository.java
@@ -13,10 +13,10 @@ import com.itwill.rest.domain.Song;
 @Repository
 public interface ArtistRoleRepository extends JpaRepository<ArtistRole, ArtistRoleId> {
 	// 음원에 참여한 아티스트 Role 객체를 리턴하는 메서드
-	List<ArtistRole> findBySongAndRoleCode_RoleId(Song song, Integer roleCodeId);
+	List<ArtistRole> findBySongAndRoleCode_RoleId(Song song, Long roleCodeId);
 
 	// 앨범에 참여한 아티스트 Role 객체를 리턴하는 메서드
-	List<ArtistRole> findBySong_AlbumAndRoleCode_RoleId(Album album, Integer roleCodeId);
+	List<ArtistRole> findBySong_AlbumAndRoleCode_RoleId(Album album, Long roleCodeId);
 
 	// 음원에 참여한 모든 아티스트 객체를 리턴하는 메서드
 	List<ArtistRole> findBySong(Song song);
