+++ rest/src/main/java/com/itwill/rest/repository/GenreCodeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itwill.rest.domain.GenreCode;
 
-public interface GenreCodeRepository extends JpaRepository<GenreCode, Integer> {
+public interface GenreCodeRepository extends JpaRepository<GenreCode, Long> {
 	
 	public GenreCode findByGenreName(String genreName);
