+++ rest/src/main/java/com/itwill/rest/repository/RoleCodeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itwill.rest.domain.RoleCode;
 
-public interface RoleCodeRepository extends JpaRepository<RoleCode, Integer> {
+public interface RoleCodeRepository extends JpaRepository<RoleCode, Long> {
 	
 }
